13.07.25 3

+++ cmd/IGym/subscription.go
@@ -6,6 +6,6 @@ type Subscription struct {
 	Sub_id       int32
 	UserSub_id   int32
 	Startdate    time.Time
-	durationdays int16
+	durationdays int32
 	enddate      time.Time
 }
